chore: move files in lana into folders
