+++ lana/src/display/Display.ts
@@ -3,7 +3,7 @@
  */
 import { Uri, commands, window } from 'vscode';
 
-import { appName } from './AppSettings.js';
+import { appName } from '../AppSettings.js';
 
 export class Display {
   private outputChannel = window.createOutputChannel(appName);
